pyproject.toml
[tool.poetry]
name = "today-hot-cn"
version = "0.1.0"
description = ""
authors = ["maguowei <imaguowei@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.0"
feedparser = "^6.0.10"
ipython = "^8.4.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
